requirements.txt
# Core dependencies
streamlit>=1.28.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.23
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0

# AI and ML
sentence-transformers>=2.2.2
transformers>=4.35.2
torch>=2.1.1
scikit-learn>=1.3.2

# LLM Integration
openai>=1.3.0  # For OpenAI GPT models
tiktoken>=0.5.1  # For token counting

# Vector storage and embeddings
chromadb>=0.4.15
# faiss-cpu>=1.7.4  # Optional fallback for vector operations

# Web scraping and async HTTP
aiohttp>=3.9.1
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Text processing and NLP
nltk>=3.8.1
spacy>=3.7.2
langdetect>=1.0.9

# Document processing
PyPDF2>=3.0.1
python-docx>=1.1.0
python-multipart>=0.0.6

# UI and visualization
plotly>=5.17.0

# Environment and utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Security (for future API features)
python-jose>=3.3.0
bcrypt>=4.1.2
